refactor: signup 페이지 폴더 구조 변경
